% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_add.R
\name{ds_add}
\alias{ds_add}
\title{ds_add}
\usage{
ds_add(ds, dataset, name, version)
}
\arguments{
\item{ds}{path to datastore}

\item{dataset}{the R-data object that should be put into the datastore}

\item{name}{name of data-object}

\item{version}{(number) version of data-object}
}
\value{
\code{TRUE} if the data-object was successfully added to the datastore
}
\description{
add a dataset to a datastore
}
\examples{
## not yet
}
