% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dataset_add.R
\name{ds_dataset_add}
\alias{ds_dataset_add}
\title{ds_dataset_add}
\usage{
ds_dataset_add(path, dataset, name, version)
}
\arguments{
\item{path}{path to datastore}

\item{dataset}{the R-data object that should be put into the datastore}

\item{name}{name of data-object}

\item{version}{(number) version of data-object}
}
\value{
\code{TRUE} if the data-object was successfully added to the datastore
}
\description{
add a dataset to a datastore
}
\examples{
## not yet
}
