% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dataset_exists.R
\name{ds_dataset_exists}
\alias{ds_dataset_exists}
\title{ds_dataset_exists}
\usage{
ds_dataset_exists(ds, name, version = NULL, verbose = TRUE)
}
\arguments{
\item{ds}{path to datastore}

\item{name}{name of data-object}

\item{version}{specific version of data-object to query; if \code{NULL}, it is checked if the object is versioned at least one time.}

\item{verbose}{(logical) if \code{TRUE}, additional messages will be printed to prompt}
}
\value{
\code{TRUE} if the data-object is available from the datastore given by argument \code{ds} or \code{FALSE} else.
}
\description{
ds_dataset_exists
}
\examples{
## not yet
}
