% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dataset_get.R
\name{ds_dataset_get}
\alias{ds_dataset_get}
\title{ds_dataset_get}
\usage{
ds_dataset_get(ds, name, version)
}
\arguments{
\item{ds}{path to datastore}

\item{name}{name of data-object to retrieve}

\item{version}{versionnumber of data-object to retrieve}
}
\value{
the requested data-object
}
\description{
retrieves a given dataset in a specific version from a datastore
}
\examples{
## not yet
}
