% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dataset_remove.R
\name{ds_dataset_remove}
\alias{ds_dataset_remove}
\title{ds_dataset_remove}
\usage{
ds_dataset_remove(path, name, version = NULL)
}
\arguments{
\item{path}{path to datastore}

\item{name}{name of data-object for which a specific version should be removed}

\item{version}{(number) version of data-object; if \code{NULL}, all versions will be removed}
}
\value{
\code{NULL}
}
\description{
ds_dataset_remove
}
\examples{
## not yet
}
