% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dataset_versions.R
\name{ds_dataset_versions}
\alias{ds_dataset_versions}
\title{ds_dataset_versions}
\usage{
ds_dataset_versions(ds, name, verbose = TRUE)
}
\arguments{
\item{ds}{path to datastore}

\item{name}{name of data-object}

\item{verbose}{(logical) if TRUE, additional messages will be printed to prompt}
}
\value{
a \code{data.frame} or \code{NULL}
}
\description{
returns a data.frame (or \code{NULL}) containing information on all versions of the requested data-object given by argument \code{name}
}
\examples{
## not yet
}
