% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_delete_datastore.R
\name{ds_delete_datastore}
\alias{ds_delete_datastore}
\title{ds_delete_datastore}
\usage{
ds_delete_datastore(ds)
}
\arguments{
\item{ds}{path to datastore}
}
\value{
\code{NULL}
}
\description{
deletes a datastore and all files within it
}
\examples{
## not yet
}
