% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_get.R
\name{ds_get}
\alias{ds_get}
\title{ds_get}
\usage{
ds_get(ds, ds_name, version)
}
\arguments{
\item{ds}{path to datastore}

\item{ds_name}{name of data-object to retrieve}

\item{version}{versionnumber of data-object to retrieve}
}
\value{
the requested data-object
}
\description{
retrieves a given dataset in a specific version from a datastore
}
\examples{
## not yet
}
