% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_get_latest.R
\name{ds_get_latest}
\alias{ds_get_latest}
\title{ds_get_latest}
\usage{
ds_get_latest(ds, ds_name)
}
\arguments{
\item{ds}{path to datastore}

\item{ds_name}{name of data-object to retrieve}
}
\value{
the requested data-object
}
\description{
retrieves the latest version of a given dataset from a datastore
}
\examples{
## not yet
}
