% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_info.R
\name{ds_info}
\alias{ds_info}
\title{ds_info}
\usage{
ds_info(ds, verbose = TRUE)
}
\arguments{
\item{ds}{path to datastore}

\item{verbose}{(logical) if \code{TRUE}, additional messages will be printed to prompt}
}
\value{
a list with the following elements
\itemize{
\item \strong{ds_name}: name of the datastore
\item \strong{created_at}: timestamp when the datastare has been created
\item \strong{datastore_path}: path to the root of the datastore
\item \strong{datasets}: names of data objects for which at least one version is stored in the datastore\code{NA}
\item \strong{nr_datasets}: number of different data objects available from the datastore
\item \strong{nr_files}: number of datafiles stored
}
}
\description{
ds_info
}
\examples{
## not yet
}
