% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_new.R
\name{ds_new}
\alias{ds_new}
\title{ds_new}
\usage{
ds_new(path)
}
\arguments{
\item{path}{Destination download path. This path is the root folder of your new repository.}
}
\description{
This function creates a new local datastore at a specified destination folder
and creates the DATASETS index file.
}
\section{datastore folder structure}{

The folder structure of a repository
\itemize{
 \item{Root}
 \itemize{
   \item{datasets/files}
   \itemize{
     \item{DATASETS}
   }
 }
}
}

\examples{
## not yet
}
