% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_read_index.R
\name{ds_read_index}
\alias{ds_read_index}
\title{ds_read_index}
\usage{
ds_read_index(path)
}
\arguments{
\item{path}{path to datastore}
}
\value{
a list containing the following elements
\itemize{
\item \strong{last_indexed}: timestamp
\item \strong{datasets}: the name of the datasets that are versioned
\item \strong{info}: a \code{data.frame} containing information about all versioned data-objects
}
}
\description{
reads (json)-index file about all versioned data-objects from specified datastore
}
\examples{
## not yet
}
