% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_update_index.R
\name{ds_update_index}
\alias{ds_update_index}
\title{ds_update_index}
\usage{
ds_update_index(path, verbose = TRUE)
}
\arguments{
\item{path}{path to datastore}

\item{verbose}{(logical) if \code{TRUE}, additional messages will be printed to prompt}
}
\value{
\code{NULL}
}
\description{
updates index file (json-format) of a given datastore
}
\examples{
## not yet
}
