% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dataset_remove.R
\name{ds_dataset_remove}
\alias{ds_dataset_remove}
\title{ds_dataset_remove}
\usage{
ds_dataset_remove(ds, name, version = NULL, verbose = TRUE)
}
\arguments{
\item{ds}{path to datastore}

\item{name}{name of data-object for which a specific version should be removed}

\item{version}{(number) version of data-object; if \code{NULL}, all versions will be removed}

\item{verbose}{(logical) if \code{TRUE}, additional messages will be printed to prompt}
}
\value{
\code{NULL}
}
\description{
ds_dataset_remove
}
\examples{
## not yet
}
